Add requests authorization

+++ src/controllers/card.ts
@@ -3,6 +3,7 @@ import { ObjectId, isValidObjectId } from 'mongoose';
 
 import ClientError from '../errors/client';
 import NotFoundError from '../errors/notFound';
+import { IRequestWithUserID } from '../middlewares/auth';
 import Card, { ICard } from '../models/card';
 
 type TCreateCard = Pick<ICard, 'name' | 'link'>;
@@ -30,15 +31,14 @@ export const createCard = async (
   next: NextFunction,
 ) => {
   try {
-    // @ts-ignore
-    const { _id } = req.user;
+    const id = (req as IRequestWithUserID).user?.id;
     const { name, link } = req.body;
 
     if (!name || !link) {
       throw new ClientError('Некорректные данные');
     }
 
-    const card = await Card.create({ name, link, owner: _id });
+    const card = await Card.create({ name, link, owner: id });
 
     res.status(201).send({ status: 'success', card });
   } catch (error) {
@@ -49,12 +49,13 @@ export const createCard = async (
 export const deleteCard = async (req: Request<ICardId>, res: Response, next: NextFunction) => {
   try {
     const { cardId } = req.params;
+    const id = (req as unknown as IRequestWithUserID).user?.id;
 
     if (!isValidObjectId(cardId)) {
       throw new ClientError(INCORRECT_ID_TEXT);
     }
 
-    const result = await Card.deleteOne({ _id: cardId });
+    const result = await Card.deleteOne({ _id: cardId, owner: id });
 
     if (result.deletedCount === 0) {
       throw new NotFoundError(NOT_FOUND_TEXT);
